Refine DynamicServiceAutoConfigurationComponentTest to not pollute spring context
with configuration classes that fail other SCAB tests: inlined them for now

CreateInstanceWithPropertiesComponentTest fails to start with message:
```
Caused by: org.springframework.beans.factory.support.BeanDefinitionOverrideException: Invalid bean definition with name 'cloudFoundryDeploymentProperties' defined in class path resource [org/springframework/cloud/appbroker/autoconfigure/CloudFoundryAppDeployerAutoConfiguration.class]: Cannot register bean definition [Root bean: class [null]; scope=; abstract=false; lazyInit=false; autowireMode=3; dependencyCheck=0; autowireCandidate=true; primary=false; factoryBeanName=org.springframework.cloud.appbroker.autoconfigure.CloudFoundryAppDeployerAutoConfiguration; factoryMethodName=cloudFoundryDeploymentProperties; initMethodName=null; destroyMethodName=(inferred); defined in class path resource [org/springframework/cloud/appbroker/autoconfigure/CloudFoundryAppDeployerAutoConfiguration.class]] for bean 'cloudFoundryDeploymentProperties': There is already [Root bean: class [null]; scope=; abstract=false; lazyInit=false; autowireMode=3; dependencyCheck=0; autowireCandidate=true; primary=false; factoryBeanName=targetPropertiesConfiguration; factoryMethodName=cloudFoundryDeploymentProperties; initMethodName=null; destroyMethodName=(inferred); defined in class path resource [org/springframework/cloud/appbroker/integration/fixtures/TargetPropertiesConfiguration.class]] bound.
```

Suspecting that this is because through its super class WiremockComponentTest, it starts AppBrokerApplication which attempts to declare a bean which collides due to our pollution
