+++ spring-cloud-app-broker-integration-tests/src/test/java/org.springframework.cloud.appbroker/integration/DynamicServiceAutoConfigurationComponentTest.java
@@ -1,12 +1,28 @@
 package org.springframework.cloud.appbroker.integration;
 
+import java.net.URI;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jdk.nashorn.internal.ir.annotations.Ignore;
+import org.cloudfoundry.client.CloudFoundryClient;
+import org.cloudfoundry.doppler.DopplerClient;
+import org.cloudfoundry.operations.CloudFoundryOperations;
+import org.cloudfoundry.operations.DefaultCloudFoundryOperations;
+import org.cloudfoundry.reactor.ConnectionContext;
+import org.cloudfoundry.reactor.DefaultConnectionContext;
+import org.cloudfoundry.reactor.ProxyConfiguration;
+import org.cloudfoundry.reactor.TokenProvider;
+import org.cloudfoundry.reactor.client.ReactorCloudFoundryClient;
+import org.cloudfoundry.reactor.doppler.ReactorDopplerClient;
+import org.cloudfoundry.reactor.tokenprovider.ClientCredentialsGrantTokenProvider;
+import org.cloudfoundry.reactor.tokenprovider.PasswordGrantTokenProvider;
+import org.cloudfoundry.reactor.uaa.ReactorUaaClient;
+import org.cloudfoundry.uaa.UaaClient;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
@@ -18,24 +34,31 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.cloud.appbroker.autoconfigure.DynamicCatalogConstants;
 import org.springframework.cloud.appbroker.autoconfigure.DynamicCatalogServiceAutoConfiguration;
 import org.springframework.cloud.appbroker.autoconfigure.ServiceDefinitionMapperProperties;
 import org.springframework.cloud.appbroker.deployer.BrokeredServices;
+import org.springframework.cloud.appbroker.deployer.cloudfoundry.CloudFoundryDeploymentProperties;
+import org.springframework.cloud.appbroker.deployer.cloudfoundry.CloudFoundryTargetProperties;
 import org.springframework.cloud.appbroker.integration.fixtures.CloudControllerStubFixture;
-import org.springframework.cloud.appbroker.integration.fixtures.CloudFoundryClientConfiguration;
+
 import org.springframework.cloud.appbroker.integration.fixtures.CredHubStubFixture;
 import org.springframework.cloud.appbroker.integration.fixtures.ExtendedCloudControllerStubFixture;
 import org.springframework.cloud.appbroker.integration.fixtures.OpenServiceBrokerApiFixture;
-import org.springframework.cloud.appbroker.integration.fixtures.TargetPropertiesConfiguration;
 import org.springframework.cloud.appbroker.integration.fixtures.TestBindingCredentialsProviderFixture;
 import org.springframework.cloud.appbroker.integration.fixtures.UaaStubFixture;
 import org.springframework.cloud.appbroker.integration.fixtures.WiremockServerFixture;
 import org.springframework.cloud.servicebroker.model.catalog.Catalog;
 import org.springframework.cloud.servicebroker.model.catalog.Plan;
 import org.springframework.cloud.servicebroker.model.catalog.ServiceDefinition;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
@@ -277,5 +300,238 @@ class DynamicServiceAutoConfigurationComponentTest {
 	void catalogFetchingFailuresThrowsException() {
 		//TODO: fail if the flux contains error events (was the case when Jackson was not configured to ignore
 	}
+
+
+	public static class TargetPropertiesConfiguration {
+
+		//Inspired from spring-cloud-app-broker-autoconfigure/src/main/java/org/springframework/cloud/appbroker/autoconfigure/CloudFoundryAppDeployerAutoConfiguration.java
+		static final String PROPERTY_PREFIX = "spring.cloud.appbroker.acceptancetest.cloudfoundry";
+
+		@Bean
+		@ConfigurationProperties(PROPERTY_PREFIX)
+		CloudFoundryTargetProperties cloudFoundryTargetProperties() {
+			return new CloudFoundryTargetProperties();
+		}
+
+
+		//Inspired from spring-cloud-app-broker-autoconfigure/src/main/java/org/springframework/cloud/appbroker/autoconfigure/CloudFoundryAppDeployerAutoConfiguration.java
+		@Bean
+		@ConfigurationProperties(PROPERTY_PREFIX + ".properties")
+		CloudFoundryDeploymentProperties cloudFoundryDeploymentProperties() {
+			return new CloudFoundryDeploymentProperties();
+		}
+
+	}
+
+	@ConfigurationProperties(TargetPropertiesConfiguration.PROPERTY_PREFIX)
+//Inspired from spring-cloud-app-broker-acceptance-tests
+	public static class CloudFoundryProperties {
+
+		static final String PROPERTY_PREFIX = "spring.cloud.appbroker.acceptancetest.cloudfoundry";
+
+		private String apiHost;
+		private Integer apiPort;
+		private String defaultOrg;
+		private String defaultSpace;
+		private String username;
+		private String password;
+		private String clientId;
+		private String clientSecret;
+		private String identityZoneSubdomain;
+		private boolean secure = true;
+		private boolean skipSslValidation;
+
+		public String getApiHost() {
+			return apiHost;
+		}
+
+		public void setApiHost(String apiHost) {
+			this.apiHost = parseApiHost(apiHost);
+		}
+
+		public Integer getApiPort() {
+			return apiPort;
+		}
+
+		public void setApiPort(int apiPort) {
+			this.apiPort = apiPort;
+		}
+
+		public String getDefaultOrg() {
+			return defaultOrg;
+		}
+
+		public void setDefaultOrg(String defaultOrg) {
+			this.defaultOrg = defaultOrg;
+		}
+
+		public String getDefaultSpace() {
+			return defaultSpace;
+		}
+
+		public void setDefaultSpace(String defaultSpace) {
+			this.defaultSpace = defaultSpace;
+		}
+
+		public String getUsername() {
+			return username;
+		}
+
+		public void setUsername(String username) {
+			this.username = username;
+		}
+
+		public String getPassword() {
+			return password;
+		}
+
+		public void setPassword(String password) {
+			this.password = password;
+		}
+
+		public String getClientId() {
+			return clientId;
+		}
+
+		public void setClientId(String clientId) {
+			this.clientId = clientId;
+		}
+
+		public String getClientSecret() {
+			return clientSecret;
+		}
+
+		public void setClientSecret(String clientSecret) {
+			this.clientSecret = clientSecret;
+		}
+
+		public String getIdentityZoneSubdomain() {
+			return identityZoneSubdomain;
+		}
+
+		public void setIdentityZoneSubdomain(String identityZoneSubdomain) {
+			this.identityZoneSubdomain = identityZoneSubdomain;
+		}
+
+		public ProxyConfiguration getProxyConfiguration() {
+			return null;
+		}
+
+		public boolean isSecure() {
+			return secure;
+		}
+
+		public void setSecure(boolean secure) {
+			this.secure = secure;
+		}
+
+		public boolean isSkipSslValidation() {
+			return skipSslValidation;
+		}
+
+		public void setSkipSslValidation(boolean skipSslValidation) {
+			this.skipSslValidation = skipSslValidation;
+		}
+
+		private static String parseApiHost(String api) {
+			final URI uri = URI.create(api);
+			return uri.getHost() == null ? api : uri.getHost();
+		}
+
+	}
+
+	@EnableConfigurationProperties(CloudFoundryProperties.class)
+//Inspired from spring-cloud-app-broker-acceptance-tests
+	public static class CloudFoundryClientConfiguration {
+
+		public static final String APP_BROKER_CLIENT_SECRET = "app-broker-client-secret";
+		public static final String[] APP_BROKER_CLIENT_AUTHORITIES = {
+			"cloud_controller.read", "cloud_controller.write", "clients.write"
+		};
+
+		@Bean
+		CloudFoundryOperations cloudFoundryOperations(
+			CloudFoundryProperties properties,
+			CloudFoundryClient client,
+			DopplerClient dopplerClient,
+			UaaClient uaaClient) {
+			return DefaultCloudFoundryOperations.builder()
+				.cloudFoundryClient(client)
+				.dopplerClient(dopplerClient)
+				.uaaClient(uaaClient)
+				.organization(properties.getDefaultOrg())
+				.space(properties.getDefaultSpace())
+				.build();
+		}
+
+		@Bean
+		CloudFoundryClient cloudFoundryClient(ConnectionContext connectionContext,
+			@Qualifier("userCredentials") TokenProvider tokenProvider) {
+			return ReactorCloudFoundryClient.builder()
+				.connectionContext(connectionContext)
+				.tokenProvider(tokenProvider)
+				.build();
+		}
+
+		@Bean
+		ConnectionContext connectionContext(
+			CloudFoundryProperties properties) {
+			return DefaultConnectionContext.builder()
+				.apiHost(properties.getApiHost())
+				.port(Optional.ofNullable(properties.getApiPort()))
+				.skipSslValidation(properties.isSkipSslValidation())
+				.secure(properties.isSecure())
+				.build();
+		}
+
+		@Bean
+		DopplerClient dopplerClient(ConnectionContext connectionContext,
+			@Qualifier("userCredentials") TokenProvider tokenProvider) {
+			return ReactorDopplerClient.builder()
+				.connectionContext(connectionContext)
+				.tokenProvider(tokenProvider)
+				.build();
+		}
+
+		@Bean
+		UaaClient uaaClient(ConnectionContext connectionContext,
+			@Qualifier("clientCredentials") TokenProvider tokenProvider) {
+			return ReactorUaaClient.builder()
+				.connectionContext(connectionContext)
+				.tokenProvider(tokenProvider)
+				.build();
+		}
+
+		@Bean
+		@Qualifier("userCredentials")
+		@ConditionalOnProperty({
+			CloudFoundryProperties.PROPERTY_PREFIX + ".username",
+			CloudFoundryProperties.PROPERTY_PREFIX + ".password"
+		})
+		PasswordGrantTokenProvider passwordTokenProvider(
+			CloudFoundryProperties properties) {
+			return PasswordGrantTokenProvider.builder()
+				.password(properties.getPassword())
+				.username(properties.getUsername())
+				.build();
+		}
+
+		@Bean
+		@Qualifier("clientCredentials")
+		@ConditionalOnProperty({
+			CloudFoundryProperties.PROPERTY_PREFIX + ".client-id",
+			CloudFoundryProperties.PROPERTY_PREFIX + ".client-secret"
+		})
+		ClientCredentialsGrantTokenProvider clientTokenProvider(
+			CloudFoundryProperties properties) {
+			return ClientCredentialsGrantTokenProvider.builder()
+				.clientId(properties.getClientId())
+				.clientSecret(properties.getClientSecret())
+				.identityZoneSubdomain(properties.getIdentityZoneSubdomain())
+				.build();
+		}
+
+	}
+
 
 }
